REFACTOR: controller out of state
